HostInstall: Only add to default suite if we have a repo
Currently we try to add the HostInstall test to the default test suite
if the configuration has defined a scratch disk. The HostInstall
constructor causes the op-test process to exit if there is no
os_repo or os_cdrom to use as an installation source so we don't want
to invoke it unless we're sure it'll succeed.

Signed-off-by: Oliver O'Halloran <oohall@gmail.com>