HostInstall: Reword variable parsing failure messages
When deciding what tests from the HostInstall suite to set of test to
run we check for a scratch disk and a repo strings. If we don't have
one or can't parse the repo string we have a generic failure message
that is not all that helpful, so split it out into seperate failure
messages for each case.

Signed-off-by: Oliver O'Halloran <oohall@gmail.com>