Move OpTestSystemXXX to their relevant files
Having all the OpTestSystem subclasses in the OpTestSystem file doesn't
make that much sense since they have very little in common, and it makes
searching for how each system type differs pretty annoying.

Move each of the subclasses to a more relevant file and fix up all the
circular import games which probably caused this situation in the first
place.

Signed-off-by: Oliver O'Halloran <oohall@gmail.com>