logger: Re-work how console logging works
Right now there's two seperate types of log files: The "logfile" used by pexpect
to record the pexpect session output and the python logging infrastructure. This
results in the console logs being recorded seperately to the op-test logs which
makes figuring out what's actually going on very difficult at the best of times.

To try and address this switch to having the consoles write into a FileLikeLogger
which exists independently to the op-test library logger. This way we can still
record the console session to it's own log file, but we can also feed the console
output into the top-level logger.

This is a little awkward at times, but I think it's a more workable approach
overall. Time will tell I suppose.

Signed-off-by: Oliver O'Halloran <oohall@gmail.com>