console: Implement is_connected() for all consoles
Forgot to do this previously, whoops.

We might want to merge this into the base class and make it "offical" that all
consoles are pexpect based. Originally I had the idea that we might want to
implement some of the consoles with native python code, similar to what we do
with TConnection that's used to talk to the FSP. However, we're pretty
fundementally tied to the pexpect.spawn interface so dunno. Maybe just
standardizing on "all consoles are based on a command that's executed"
(ipmitool, ssh, telnet, etc) is the way to go.

For now just implement is_connected() the same way for all the derived
classes since I don't think the base class *requires* a pexpect session.

Signed-off-by: Oliver O'Halloran <oohall@gmail.com>