console: Prevent shell setup creating junk file
I noticed that after running the shell prompt setup from op-test we
sometimes get this file created:

	# ls -1
	null which sh && exec sh;
	# cat null\ which\ sh\ \&\&\ exec\ sh\;
	ENV=/dev/null which sh && exec sh;
	#

For context, the full setup is this:

	# unset HISTFILE PS1;
	ENV=/dev/null which sh && exec sh;
	/bin/sh
	~ # which stty && stty cols 300; which stty && stty rows 30;
	/bin/stty
	/bin/stty
	~ # export LANG=C;
	~ # PS1="console-expect1#"
	console-expect1#

I'm not entirely sure what's going on that causes that file to be created
and running "ENV=/dev/null which sh && exec sh;" manually doesn't trigger
the same behaviour. Wrapping "/dev/null" in quotes does seem to stop the
file being created though, so do that.

Signed-off-by: Oliver O'Halloran <oohall@gmail.com>